+++ templates/package.json.ts
@@ -1,8 +1,12 @@
-import { promiseSequence } from "alcalzone-shared/async";
 import * as JSON5 from "json5";
+import * as pLimit from "p-limit";
 import { TemplateFunction } from "../src/lib/createAdapter";
+import { licenses } from "../src/lib/licenses";
 import { fetchPackageVersion, getPackageName } from "../src/lib/packageVersions";
 import { getDefaultAnswer } from "../src/lib/questions";
+
+// Limit package version downloads to 10 simultaneous connections
+const downloadLimiter = pLimit(10);
 
 const templateFunction: TemplateFunction = async answers => {
 
@@ -19,8 +23,9 @@ const templateFunction: TemplateFunction = async answers => {
 		.concat(isAdapter ? ["@iobroker/adapter-core"] : [])
 		.sort()
 		.map((dep) => (async () => `"${dep}": "^${await fetchPackageVersion(dep, "0.0.0")}"`))
+		.map(task => downloadLimiter(task))
 		;
-	const dependencies = await promiseSequence<string>(dependencyPromises);
+	const dependencies = await Promise.all(dependencyPromises);
 
 	const devDependencyPromises = ([] as string[])
 		.concat([
@@ -98,8 +103,9 @@ const templateFunction: TemplateFunction = async answers => {
 		.concat(useNyc ? ["nyc"] : [])
 		.sort()
 		.map((dep) => (async () => `"${getPackageName(dep)}": "^${await fetchPackageVersion(dep, "0.0.0")}"`))
+		.map(task => downloadLimiter(task))
 		;
-	const devDependencies = await promiseSequence<string>(devDependencyPromises);
+	const devDependencies = await Promise.all(devDependencyPromises);
 
 	const gitUrl = answers.gitRemoteProtocol === "HTTPS"
 		? `https://github.com/${answers.authorGithub}/ioBroker.${answers.adapterName}`
@@ -120,7 +126,7 @@ const templateFunction: TemplateFunction = async answers => {
 	)},
 	`) : ""}
 	"homepage": "https://github.com/${answers.authorGithub}/ioBroker.${answers.adapterName}",
-	"license": "${answers.license!.id}",
+	"license": "${licenses[answers.license!].id}",
 	"keywords": ${JSON.stringify(answers.keywords || getDefaultAnswer("keywords"))},
 	"repository": {
 		"type": "git",
@@ -143,9 +149,9 @@ const templateFunction: TemplateFunction = async answers => {
 				${useReact ? `"watch:parcel": "parcel admin/src/index.tsx -d admin/build${useDevcontainer ? ` --hmr-port 1235` : ""}",` : ""}
 				"watch:ts": "tsc -p tsconfig.build.json --watch",
 				"watch": "npm run watch:ts",
-				"test:ts": "mocha src/**/*.test.ts",
+				"test:ts": "mocha --config test/mocharc.custom.json src/**/*.test.ts",
 			`) : (`
-				"test:js": "mocha \\"{!(node_modules|test)/**/*.test.js,*.test.js,test/**/test!(PackageFiles|Startup).js}\\"",
+				"test:js": "mocha --config test/mocharc.custom.json \\"{!(node_modules|test)/**/*.test.js,*.test.js,test/**/test!(PackageFiles|Startup).js}\\"",
 			`)}
 			"test:package": "mocha test/package --exit",
 			"test:unit": "mocha test/unit --exit",