+++ test/baselines/devcontainer/.devcontainer/parcel/run.sh
@@ -1,13 +1,7 @@
 #!/bin/bash
-echo "Starting initial sync..."
-/usr/app/sync.sh
-
-cd /usr/workspace
+cd /workspace
 
 echo "Installing all dependencies..."
 npm install
 
-# run the following two commands in parallel (honoring Ctrl-C)
-# - npm run watch:parcel
-# - /usr/app/run-sync.sh
-(echo "npm run watch:parcel"; echo "/usr/app/run-sync.sh") | xargs -I{} -n 1 -P 2 /bin/bash -c "{}"
+npm run watch:parcel