+++ test/baselines/devcontainer/.devcontainer/parcel/sync.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-rsync -aiz --inplace --exclude={'node_modules','.git','.cache','admin/build'} /workspace/ /usr/workspace